Ctrl Space - move to TitleBlockHelper + add to properties
